requirements.txt
pypdf
openai
pandas
chromadb
python-dotenv